pyproject.toml
[project]
name = "akshare-one"
version = "0.3.10"
description = "Standardized interface for Chinese financial market data, built on AKShare with unified data formats and simplified APIs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "akshare>=1.17.41",
    "cachetools>=5.5.2",
]
license = "MIT"
keywords = ["akshare", "financial-data", "stock-data", "quant"]

[project.urls]
Homepage = "https://github.com/zwldarren/akshare-one"
Repository = "https://github.com/zwldarren/akshare-one.git"

[project.optional-dependencies]
talib = [
    "ta-lib>=0.6.4",
]

[dependency-groups]
dev = [
    "mkdocs-material>=9.6.15",
    "mypy>=1.17.1",
    "pandas-stubs>=2.3.0.250703",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.1",
    "types-cachetools>=6.1.0.20250717",
    "types-requests>=2.32.4.20250809",
]

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=akshare_one --cov-report=term-missing"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.mypy]
strict = true
